tools: created credentials.py
