+++ .gitignore
@@ -1,5 +1,6 @@
 *.db
 env/
+venv/
 __pycache__/
 *.pyc
 *.pyo
